add bip137
https://github.com/RingsNetwork/rings-node/issues/410
