+++ core/Cargo.toml
@@ -60,7 +60,8 @@ futures-timer = "3.0.2"
 hex = "0.4.3"
 itertools = "0.10.3"
 libsecp256k1 = "0.7.0"
-num-bigint = "0.3.1"
+num-bigint = "0.4.3"
+num-integer = "0.1.45"
 rand = { version = "0.8.5", features = ["getrandom"] }
 rand_core = { version = "0.6.3", features = ["getrandom"] }
 rand_hc = "0.3.1"
@@ -120,6 +121,7 @@ web-sys = { version = "0.3.56", optional = true, features = [
 ] }
 
 [dev-dependencies]
+base64 = { version = "0.13.0" }
 tracing-subscriber = { version = "0.3.15", features = ["ansi"] }
 tracing-wasm = "0.2.1"
 wasm-bindgen-test = "0.3.0"
